Remove shapes.py, move code to abstract_arrays.py / api.py
