+++ jax/lax/lax.py
@@ -27,7 +27,6 @@ from typing import Any
 import warnings
 
 import numpy as onp
-from jax.interpreters.shapes import to_index, is_polymorphic, Poly
 
 from ..util import partial, prod
 
@@ -41,9 +40,9 @@ from ..config import flags
 from ..core import Primitive
 from ..abstract_arrays import (UnshapedArray, ShapedArray, ConcreteArray,
                                AbstractToken, array_types, make_shaped_array,
-                               raise_to_shaped, abstract_token)
-from ..interpreters import xla, pxla, ad, batching, shapes, masking, \
-  partial_eval as pe
+                               raise_to_shaped, abstract_token, to_index,
+                               is_polymorphic)
+from ..interpreters import xla, pxla, ad, batching, masking, partial_eval as pe
 from ..util import curry, cache, safe_zip, unzip2, prod
 from ..tree_util import build_tree, tree_unflatten, tree_map
 from ..lib import pytree
@@ -74,7 +73,7 @@ def broadcast_shapes(*shapes):
 
 def _try_canonicalize_shape(shape):
   try:
-    return tuple(map(lambda x: shapes.to_index(x), shape))
+    return tuple(map(lambda x: to_index(x), shape))
   except (TypeError, AttributeError):
     return None
