+++ tests/masking_test.py
@@ -22,8 +22,8 @@ from unittest import SkipTest
 import numpy as onp
 from absl.testing import absltest
 from jax.interpreters.masking import shape_as_value
-from jax.interpreters.shapes import ShapeError
-from jax import numpy as np, test_util as jtu, mask, vmap, jit, grad, lax
+from jax import numpy as np, test_util as jtu, mask, vmap, jit, grad, lax, \
+  ShapeError
 from jax.config import config
 config.parse_flags_with_absl()
